id 검증 로직 구현

+++ bulletin-board/src/main/java/com/solo/bulletin_board/member/service/MemberService.java
@@ -1,5 +1,6 @@
 package com.solo.bulletin_board.member.service;
 
+import com.solo.bulletin_board.auth.userDetailsService.CustomUserDetails;
 import com.solo.bulletin_board.auth.utils.CustomAuthorityUtils;
 import com.solo.bulletin_board.exception.BusinessLogicException;
 import com.solo.bulletin_board.exception.ExceptionCode;
@@ -53,6 +54,13 @@ public class MemberService {
         }
     }
 
+    public void checkMemberId(long memberId, CustomUserDetails customUserDetails){
+
+        if(memberId != customUserDetails.getMemberId()){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_MATCHED);
+        }
+    }
+
     public Member createMember(Member member){
 
         verifyExistsMember(member.getEmail());
@@ -68,9 +76,9 @@ public class MemberService {
 
     }
 
-    public Member updateMember(Member member){
+    public Member updateMember(Member member, CustomUserDetails customUserDetails){
 
-        Member findMember = findVerifiedMember(member.getMemberId());
+        Member findMember = findVerifiedMember(customUserDetails.getMemberId());
 
         Optional.ofNullable(member.getNickname())
                 .ifPresent(nickname -> {
@@ -84,14 +92,19 @@ public class MemberService {
 
     }
 
-    public Member findMember(long memberId){
+    public Member findMember(CustomUserDetails customUserDetails){
 
-        return findVerifiedMember(memberId);
+        return findVerifiedMember(customUserDetails.getMemberId());
     }
 
-    public void deleteMember(long memberId){
+    public void deleteMember(String password, CustomUserDetails customUserDetails){
 
-        Member findMember = findVerifiedMember(memberId);
+        Member findMember = findVerifiedMember(customUserDetails.getMemberId());
+
+        if(!passwordEncoder.matches(password, findMember.getPassword())){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_PASSWORD_NOT_MATCHED);
+        }
+
         memberRepository.delete(findMember);
     }
